Define and use constants for opcode numbers.
Hard coding the opcode numbers in the function just makes it harder to figure
out what's going on, but much more to the point, not defining the opcodes in
the header makes it impossible to use the generated headers instead of the
x11proto headers in the server.

The name I settled on is very simple, for an extension by the name of xconf,
and a request by the name of list_devices, we get XCB_XCONF_LIST_DEVICES.  If
this somehow causes problems, we can probably add a _OP somewhere in there,
but.

Acked-by: Jamey Sharp <jamey@minilop.net>
Closes: #8641