xcb_send_fd(): Always close fds
The API docs for xcb_send_fd() says "After this function returns, the file
descriptor given is owned by xcb and will be closed eventually".

Let the implementation live up to its documentation. We now also close fds if fd
passing is unavailable (!HAVE_SENDMSG) and when the connection is in an error
state.

(This also does sneak in some preparatory functions for follow-up commits and
thus does things in a more complicated way than really necessary.)

Signed-off-by: Uli Schlachter <psychon@znc.in>